requirements.txt
# Core Django packages
django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.9
dj-database-url==0.5.0

# Environment and settings
python-decouple==3.8
redis==5.0.0
whitenoise==6.5.0
django-cors-headers==4.3.0
